requirements.txt
decorator==4.0.2
Django==1.7.7
django-crispy-forms==1.5.0
django-user-roles==0.1.0
djangorestframework==3.2.2
hamlpy==0.82.2
numpy==1.9.2
python-social-auth==0.2.12
requests==2.7.0
requests-oauthlib==0.5.0
psycopg2
expfactory
boto==2.1.1
django-dbbackup
cognitiveatlas
dropbox
kombu
opbeat
djrill
